+++ devito/__init__.py
@@ -1,0 +1,3 @@
+class NamedObject(object):
+    def getName(self):
+        return self.__class__.__name__
